+++ src/rpc/server.cpp
@@ -35,6 +35,32 @@ static RPCTimerInterface *timerInterface = nullptr;
 /* Map of name to timer. */
 static std::map<std::string, std::unique_ptr<RPCTimerBase>> deadlineTimers;
 
+struct RPCCommandExecutionInfo {
+    std::string method;
+    int64_t start;
+};
+
+struct RPCServerInfo {
+    Mutex mutex;
+    std::list<RPCCommandExecutionInfo> active_commands GUARDED_BY(mutex);
+};
+
+static RPCServerInfo g_rpc_server_info;
+
+struct RPCCommandExecution {
+    std::list<RPCCommandExecutionInfo>::iterator it;
+    explicit RPCCommandExecution(const std::string &method) {
+        LOCK(g_rpc_server_info.mutex);
+        it = g_rpc_server_info.active_commands.insert(
+            g_rpc_server_info.active_commands.cend(),
+            {method, GetTimeMicros()});
+    }
+    ~RPCCommandExecution() {
+        LOCK(g_rpc_server_info.mutex);
+        g_rpc_server_info.active_commands.erase(it);
+    }
+};
+
 UniValue RPCServer::ExecuteCommand(Config &config,
                                    const JSONRPCRequest &request) const {
     // Return immediately if in warmup
@@ -75,7 +101,6 @@ void RPCServer::RegisterCommand(std::unique_ptr<RPCCommand> command) {
 static struct CRPCSignals {
     boost::signals2::signal<void()> Started;
     boost::signals2::signal<void()> Stopped;
-    boost::signals2::signal<void(const ContextFreeRPCCommand &)> PreCommand;
 } g_rpcSignals;
 
 void RPCServerSignals::OnStarted(std::function<void()> slot) {
@@ -319,6 +344,29 @@ static UniValue uptime(const Config &config,
     return GetTime() - GetStartupTime();
 }
 
+static UniValue getrpcinfo(const Config &config,
+                           const JSONRPCRequest &request) {
+    if (request.fHelp || request.params.size() > 0) {
+        throw std::runtime_error("getrpcinfo\n"
+                                 "\nReturns details of the RPC server.\n");
+    }
+
+    LOCK(g_rpc_server_info.mutex);
+    UniValue active_commands(UniValue::VARR);
+    for (const RPCCommandExecutionInfo &info :
+         g_rpc_server_info.active_commands) {
+        UniValue entry(UniValue::VOBJ);
+        entry.pushKV("method", info.method);
+        entry.pushKV("duration", GetTimeMicros() - info.start);
+        active_commands.push_back(entry);
+    }
+
+    UniValue result(UniValue::VOBJ);
+    result.pushKV("active_commands", active_commands);
+
+    return result;
+}
+
 /**
  * Call Table
  */
@@ -327,6 +375,7 @@ static const ContextFreeRPCCommand vRPCCommands[] = {
     //  category            name                      actor (function)        argNames
     //  ------------------- ------------------------  ----------------------  ----------
     /* Overall control/query calls */
+    { "control",            "getrpcinfo",             getrpcinfo,             {}  },
     { "control",            "help",                   help,                   {"command"}  },
     { "control",            "stop",                   stop,                   {"wait"}  },
     { "control",            "uptime",                 uptime,                 {}  },
@@ -521,9 +570,8 @@ UniValue CRPCTable::execute(Config &config,
         throw JSONRPCError(RPC_METHOD_NOT_FOUND, "Method not found");
     }
 
-    g_rpcSignals.PreCommand(*pcmd);
-
     try {
+        RPCCommandExecution execution(request.strMethod);
         // Execute, convert arguments to array if necessary
         if (request.params.isObject()) {
             return pcmd->call(config,